reader: detect UnexpectedEOF, split on Content-Length
If the end of the input stream is reached during the WARC headers, this
is obviously a malformed stream and should be reported as an unexpected
EOF.

Additionally, if an error occurs while scanning that error should be
returned to the caller!

When the end of the WARC headers are reached, we check for a
Content-Length header and use that to read the full record body - some
record bodies, such as the end of HTTP headers or random chance in
compressed files, will contain \r\n\r\n so a double newline is not an
appropriate place to stop reading the content body.

The content-length body reading is not applied for segmented records,
simply because I don't have any of those to test the code on / read to
see if the double newline splitting behavior is correct for those.